Extract to auditor-test-utils
